[BE] Refactor: auth 관련 미들웨어 middlewares/auth.ts 파일로 분리
- autheticateJWT, checkAuth 미들웨어 분리
- auth 라우터 컨트롤러 함수 분리
- checkAuth에서 user 없을 시 403 에러 처리
